delete stepper and move into register page
